+++ packages/assignment/src/contexts/themeContext.ts
@@ -1,0 +1,12 @@
+import { createContext, useContext } from 'react';
+import { ThemeContextType } from '../types';
+
+export const ThemeContext = createContext<ThemeContextType | undefined>(undefined);
+
+export const useThemeContext = () => {
+  const context = useContext(ThemeContext);
+  if (context === undefined) {
+    throw new Error('useTheme must be used within a ThemeProvider');
+  }
+  return context;
+};
