+++ packages/assignment/src/providers/NotificationProvider.tsx
@@ -1,0 +1,32 @@
+import { PropsWithChildren, useState, useCallback } from 'react';
+import { useMemo } from '../@lib';
+import { NotificationContext } from '../contexts';
+import { Notification } from '../types';
+
+export const NotificationProvider: React.FC<PropsWithChildren> = ({ children }) => {
+  const [notifications, setNotifications] = useState<Notification[]>([]);
+
+  const addNotification = useCallback((message: string, type: Notification['type']) => {
+    const newNotification: Notification = {
+      id: Date.now(),
+      message,
+      type,
+    };
+    setNotifications((prev) => [...prev, newNotification]);
+  }, []);
+
+  const removeNotification = useCallback((id: number) => {
+    setNotifications((prev) => prev.filter((notification) => notification.id !== id));
+  }, []);
+
+  const value = useMemo(
+    () => ({
+      notifications,
+      addNotification,
+      removeNotification,
+    }),
+    [notifications, addNotification, removeNotification] // 이제 addNotification이 메모이제이션됨
+  );
+
+  return <NotificationContext.Provider value={value}>{children}</NotificationContext.Provider>;
+};
