+++ packages/assignment/src/providers/index.tsx
@@ -1,0 +1,14 @@
+import { PropsWithChildren } from 'react';
+import { ThemeProvider } from './ThemeProvider';
+import { AuthProvider } from './AuthProvider';
+import { NotificationProvider } from './NotificationProvider';
+
+export const RootProviders: React.FC<PropsWithChildren> = ({ children }) => {
+  return (
+    <ThemeProvider>
+      <NotificationProvider>
+        <AuthProvider>{children}</AuthProvider>
+      </NotificationProvider>
+    </ThemeProvider>
+  );
+};
